Refactor SafeSearchSettingsActivity and SafeSearchListAdapter
